fix: global 패키지 리팩터링
