+++ src/main/java/or/sopt/assignment/domain/post/service/PostService.java
@@ -9,11 +9,10 @@ import or.sopt.assignment.domain.post.controller.dto.PostUpdateRequestDTO;
 import or.sopt.assignment.domain.post.entity.Post;
 import or.sopt.assignment.domain.post.repository.PostRepository;
 import or.sopt.assignment.domain.user.entity.User;
-import or.sopt.assignment.global.api.status.CommonErrorStatus;
-import or.sopt.assignment.global.exception.handler.PostHandler;
-import or.sopt.assignment.global.exception.handler.UserHandler;
-import or.sopt.assignment.global.port.LocalDateTime;
-import or.sopt.assignment.global.api.status.ErrorStatus;
+import or.sopt.assignment.global.api.exception.status.CommonErrorStatus;
+import or.sopt.assignment.global.api.exception.handler.PostHandler;
+import or.sopt.assignment.global.api.exception.handler.UserHandler;
+import or.sopt.assignment.global.infrastructure.LocalDateTime;
 import or.sopt.assignment.domain.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
