+++ src/main/java/or/sopt/assignment/global/api/exception/status/CommonErrorStatus.java
@@ -1,4 +1,4 @@
-package or.sopt.assignment.global.api.status;
+package or.sopt.assignment.global.api.exception.status;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
