+++ src/main/java/or/sopt/assignment/global/auth/jwt/JWTUtil.java
@@ -1,4 +1,4 @@
-package or.sopt.assignment.global.jwt;
+package or.sopt.assignment.global.auth.jwt;
 
 public interface JWTUtil {
     Long getId(String token);
