+++ src/main/java/or/sopt/assignment/global/auth/jwt/JWTUtilImpl.java
@@ -1,4 +1,4 @@
-package or.sopt.assignment.global.jwt;
+package or.sopt.assignment.global.auth.jwt;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
