+++ src/main/java/or/sopt/assignment/global/auth/jwt/exception/TokenHandler.java
@@ -1,0 +1,10 @@
+package or.sopt.assignment.global.auth.jwt.exception;
+
+import or.sopt.assignment.global.api.exception.status.ErrorStatus;
+import or.sopt.assignment.global.api.exception.CustomException;
+
+public class TokenHandler extends CustomException {
+    public TokenHandler(ErrorStatus errorStatus) {
+        super(errorStatus);
+    }
+}
