--- src/main/java/or/sopt/assignment/global/exception/handler/PostHandler.java
@@ -1,13 +1,0 @@
-package or.sopt.assignment.global.exception.handler;
-
-import or.sopt.assignment.global.exception.CustomException;
-import or.sopt.assignment.global.api.status.ErrorStatus;
-
-public class PostHandler extends CustomException {
-
-    public PostHandler(ErrorStatus errorDTO) {
-
-        super(errorDTO);
-
-    }
-}
