--- src/main/java/or/sopt/assignment/global/exception/handler/UserHandler.java
@@ -1,10 +1,0 @@
-package or.sopt.assignment.global.exception.handler;
-
-import or.sopt.assignment.global.exception.CustomException;
-import or.sopt.assignment.global.api.status.ErrorStatus;
-
-public class UserHandler extends CustomException {
-    public UserHandler(ErrorStatus errorStatus) {
-        super(errorStatus);
-    }
-}
